Billboards.
Made Vector3 a case class.  RenderPortal uses billboarding rendering just for test.  Billboard uses the perpendicular-to-view-frustum approach, rather than Looks-At-Camera-Loc.  This is cause that's how minecraft particles do it, and the code is basically ripped directly from that.

Rendering billboards is done with directly-passed-in locations + transformations, to TileEntitySpecialRenderers, and things.  No extra translations required.